fix/allocation-mutation
