+++ api/src/schema.gql
@@ -208,14 +208,13 @@ type Mutation {
   removeList(removeList: RemoveListInput): Boolean!
   updateListSort(updateListSort: UpdateListSort!): Boolean!
   updateTaskSort(updateTask: UpdateTaskSort!): [Task!]!
-  addTask(newTask: NewTaskInput!): Task!
+  addTask(assignTask: assignTaskInput!, newTask: NewTaskInput!): Task!
   updateTaskTitle(taskId: Float!, title: String!): Task!
   updateTaskOverview(taskId: Float!, overview: String!): Task!
   updateTaskParameters(taskId: Float!, design: Float!, plan: Float!, motivation: Float!, solution: Float!, achievement: Float!, technology: Float!): Task!
   updateTaskEndDate(taskId: Float!, end_date: String!): Task!
   deleteTask(taskId: Float!): [Task!]!
   createAllocation(newAllocation: NewAllocationInput!): Allocation!
-  assignTask(assignTask: assignTaskInput!): [Allocation!]!
   unassignTask(task_id: Float!, user_id: String!): [Allocation!]!
   addChat(userId: String!, taskId: Float!, comment: String!): Chat!
   updateChat(taskId: Float!, comment: String!, chatId: Float!): [Chat!]!
@@ -319,6 +318,14 @@ input UpdateTask {
   completed_flg: Boolean!
 }
 
+input assignTaskInput {
+  users: [assignUserInput!]!
+}
+
+input assignUserInput {
+  user_id: String!
+}
+
 input NewTaskInput {
   title: String!
   overview: String!
@@ -338,15 +345,6 @@ input NewTaskInput {
 input NewAllocationInput {
   user_id: String!
   task_id: String!
-}
-
-input assignTaskInput {
-  users: [assignUserInput!]!
-  task_id: String!
-}
-
-input assignUserInput {
-  user_id: String!
 }
 
 type Subscription {
