+++ api/src/tasks/tasks.resolver.ts
@@ -1,4 +1,5 @@
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { assignTaskInput } from 'src/allocations/dto/newAllocation.input';
 import { NewTaskInput } from './dto/newTask.input';
 import { UpdateTaskSort } from './dto/updateTaskSort.input';
 import { Task } from './task';
@@ -18,8 +19,11 @@ export class TasksResolver {
   }
 
   @Mutation(() => Task)
-  public async addTask(@Args({ name: 'newTask' }) newTask: NewTaskInput) {
-    return await this.taskService.addTask(newTask).catch((err) => {
+  public async addTask(
+    @Args({ name: 'newTask' }) newTask: NewTaskInput,
+    @Args({ name: 'assignTask' }) assignTask: assignTaskInput,
+  ) {
+    return await this.taskService.addTask(newTask, assignTask).catch((err) => {
       throw err;
     });
   }
